Split Classes into files and clean up
