+++ src/Directions.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class Directions(Enum):
+    UP = "up"
+    RIGHT = "right"
+    DOWN = "down"
+    LEFT = "left"
