+++ src/Point.py
@@ -1,0 +1,4 @@
+class Point():
+    def __init__(self, x: int, y: int):
+        self.x = x
+        self.y = y
