+++ src/config.py
@@ -1,0 +1,15 @@
+#RGB Colors
+WHITE = 255, 255, 255
+BLACK = 0, 0, 0
+RED = 255, 0, 0
+
+#Game settings
+POINT_SIZE = 80
+
+#Window settings
+WINDOW_SIZE = width, height = 1280, 720
+WINDOW_NAME = "AI Snake"
+MIDDLE_X, MIDDLE_Y = int(width / 2) - POINT_SIZE, int(height / 2) - POINT_SIZE
+FPS = 10
+
+
